+++ changes.md
@@ -1,10 +1,14 @@
+# Changes for current commit
 ## Changes in Backend Code
 ### music_classes.py
-- replaced (new) SpotifyID/URI Types with (old) str Types -> fewer errors and conversions 
-- added Popularity and Blacklisted to every class including Genre (new)
-- Album, Artist, Track and User can be initialized without errors; Playlist WIP; other not tested yet
+- fixed Adding Playlist/Album ID to other Types -> Playlist() works
+- Player() can be called even if Spotify is not running
+
+### database_access.py
+- added Device()/devices to Database
 
 ### other files
 - minor updates/changes
 
 ## Changes in Frontend Code
+- none